refactor: change types/review-section to types/review
