+++ client/components/review-section/edit-delete-review.tsx
@@ -2,7 +2,7 @@ import { Button } from '@/components/ui/button'
 import { useUser } from '@/lib/auth'
 import { deleteReview } from '@/lib/request'
 import { cn } from '@/lib/utils'
-import { Review } from '@/types/review-section'
+import { Review } from '@/types/review'
 import { Pencil1Icon, TrashIcon } from '@radix-ui/react-icons'
 import { useSetAtom } from 'jotai'
 import { toast } from 'sonner'
