+++ client/components/review-section/upvote-button.tsx
@@ -4,7 +4,7 @@ import { ArrowUpCircleIcon, ArrowUpCircleSolidIcon } from '@/components/icons'
 import { Button } from '@/components/ui/button'
 import { getUserData } from '@/lib/auth'
 import { upvote } from '@/lib/request'
-import { Review } from '@/types/review-section'
+import { Review } from '@/types/review'
 import { useAtom } from 'jotai'
 import { toast } from 'sonner'
 import {
